requirements-gold.txt
delta-spark==3.2.0
google-cloud-storage
pyspark[gcp]